requirements.txt
crewai
crewai-tools
langchain-openai
python-dotenv
fastapi
uvicorn
pydantic
httpx
supabase
gotrue